Improve agent getting started samples
