+++ python/samples/getting_started_with_agents/azure_ai_agent/step1_azure_ai_agent.py
@@ -11,9 +11,9 @@ from semantic_kernel.contents.utils.author_role import AuthorRole
 from semantic_kernel.functions.kernel_function_decorator import kernel_function
 
 ###################################################################
-# The following sample demonstrates how to create a simple,       #
-# Azure AI agent that answers questions about a sample menu       #
-# using a Semantic Kernel Plugin.                                 #
+# The following sample demonstrates how to create an Azure AI     #
+# agent that answers questions about a sample menu using a        #
+# Semantic Kernel Plugin.                                         #
 ###################################################################
 
 
@@ -36,8 +36,18 @@ class MenuPlugin:
         return "$9.99"
 
 
+# Simulate a conversation with the agent
+USER_INPUTS = [
+    "Hello",
+    "What is the special soup?",
+    "How much does that cost?",
+    "Thank you",
+]
+
+
 async def main() -> None:
     ai_agent_settings = AzureAIAgentSettings.create()
+    assert ai_agent_settings.project_connection_string, "Please provide a valid Azure AI connection string."  # nosec
 
     async with (
         DefaultAzureCredential() as creds,
@@ -46,17 +56,14 @@ async def main() -> None:
             conn_str=ai_agent_settings.project_connection_string.get_secret_value(),
         ) as client,
     ):
-        AGENT_NAME = "Host"
-        AGENT_INSTRUCTIONS = "Answer questions about the menu."
-
-        # Create agent definition
+        # 1. Create an agent on the Azure AI agent service
         agent_definition = await client.agents.create_agent(
             model=ai_agent_settings.model_deployment_name,
-            name=AGENT_NAME,
-            instructions=AGENT_INSTRUCTIONS,
+            name="Host",
+            instructions="Answer questions about the menu.",
         )
 
-        # Create the AzureAI Agent
+        # 2. Create a Semantic Kernel agent for the Azure AI agent
         agent = AzureAIAgent(
             client=client,
             definition=agent_definition,
@@ -64,27 +71,20 @@ async def main() -> None:
             # polling_options=RunPollingOptions(run_polling_interval=timedelta(seconds=1)),
         )
 
-        # Add the sample plugin to the kernel
+        # 3. Add a plugin to the agent via the kernel
         agent.kernel.add_plugin(MenuPlugin(), plugin_name="menu")
 
-        # Create a new thread
+        # 4. Create a new thread on the Azure AI agent service
         thread = await client.agents.create_thread()
 
-        user_inputs = [
-            "Hello",
-            "What is the special soup?",
-            "How much does that cost?",
-            "Thank you",
-        ]
-
         try:
-            for user_input in user_inputs:
-                # Add the user input as a chat message
+            for user_input in USER_INPUTS:
+                # 5. Add the user input as a chat message
                 await agent.add_chat_message(
                     thread_id=thread.id, message=ChatMessageContent(role=AuthorRole.USER, content=user_input)
                 )
                 print(f"# User: '{user_input}'")
-                # Invoke the agent for the specified thread
+                # 6. Invoke the agent for the specified thread for response
                 async for content in agent.invoke(
                     thread_id=thread.id,
                     temperature=0.2,  # override the agent-level temperature setting with a run-time value
@@ -92,8 +92,15 @@ async def main() -> None:
                     if content.role != AuthorRole.TOOL:
                         print(f"# Agent: {content.content}")
         finally:
+            # 7. Cleanup: Delete the thread and agent
             await client.agents.delete_thread(thread.id)
             await client.agents.delete_agent(agent.id)
+
+        # Sample Output:
+        # User: 'Hello'
+        # Agent: Hello! How can I assist you today?
+        # User: 'What is the special soup?'
+        # ...
 
 
 if __name__ == "__main__":
