+++ python/samples/getting_started_with_agents/chat_completion/README.md
@@ -1,3 +1,3 @@
-## Chat Completion Agents
+# Chat Completion Agents
 
 The following getting started samples show how to use Chat Completion agents with Semantic Kernel.
