+++ python/samples/getting_started_with_agents/chat_completion/step1_chat_completion_agent.py
@@ -1,0 +1,101 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+import asyncio
+from typing import Annotated
+
+from semantic_kernel import Kernel
+from semantic_kernel.agents import ChatCompletionAgent
+from semantic_kernel.connectors.ai import FunctionChoiceBehavior
+from semantic_kernel.connectors.ai.open_ai import AzureChatCompletion
+from semantic_kernel.contents import ChatHistory
+from semantic_kernel.contents.function_call_content import FunctionCallContent
+from semantic_kernel.contents.function_result_content import FunctionResultContent
+from semantic_kernel.functions import KernelArguments, kernel_function
+
+###################################################################
+# The following sample demonstrates how to create a chat          #
+# completion agent that answers questions about a sample menu     #
+# using a Semantic Kernel Plugin.                                 #
+###################################################################
+
+
+# Define a sample plugin for the sample
+class MenuPlugin:
+    """A sample Menu Plugin used for the concept sample."""
+
+    @kernel_function(description="Provides a list of specials from the menu.")
+    def get_specials(self) -> Annotated[str, "Returns the specials from the menu."]:
+        return """
+        Special Soup: Clam Chowder
+        Special Salad: Cobb Salad
+        Special Drink: Chai Tea
+        """
+
+    @kernel_function(description="Provides the price of the requested menu item.")
+    def get_item_price(
+        self, menu_item: Annotated[str, "The name of the menu item."]
+    ) -> Annotated[str, "Returns the price of the menu item."]:
+        return "$9.99"
+
+
+# Simulate a conversation with the agent
+USER_INPUTS = [
+    "Hello",
+    "What is the special soup?",
+    "What does that cost?",
+    "Thank you",
+]
+
+
+async def main():
+    # 1. Create the instance of the Kernel to register the plugin and service
+    service_id = "agent"
+    kernel = Kernel()
+    kernel.add_plugin(MenuPlugin(), plugin_name="menu")
+    kernel.add_service(AzureChatCompletion(service_id=service_id))
+
+    # 2. Configure the function choice behavior to auto invoke kernel functions
+    # so that the agent can automatically execute the menu plugin functions when needed
+    settings = kernel.get_prompt_execution_settings_from_service_id(service_id=service_id)
+    settings.function_choice_behavior = FunctionChoiceBehavior.Auto()
+
+    # 3. Create the agent
+    agent = ChatCompletionAgent(
+        service_id=service_id,
+        kernel=kernel,
+        name="Host",
+        instructions="Answer questions about the menu.",
+        arguments=KernelArguments(settings=settings),
+    )
+
+    # 4. Create a chat history to hold the conversation
+    chat_history = ChatHistory()
+
+    for user_input in USER_INPUTS:
+        # 5. Add the user input to the chat history
+        chat_history.add_user_message(user_input)
+        print(f"# User: {user_input}")
+        # 6. Invoke the agent for a response
+        async for content in agent.invoke(chat_history):
+            print(f"# {content.name}: ", end="")
+            if (
+                not any(isinstance(item, (FunctionCallContent, FunctionResultContent)) for item in content.items)
+                and content.content.strip()
+            ):
+                # We only want to print the content if it's not a function call or result
+                print(f"{content.content}", end="", flush=True)
+        print("")
+
+    # Sample output:
+    # User: Hello
+    # Host: Hello! How can I assist you today?
+    # User: What is the special soup?
+    # Host: The special soup is Clam Chowder.
+    # User: What does that cost?
+    # Host: The special soup, Clam Chowder, costs $9.99.
+    # User: Thank you
+    # Host: You're welcome! If you have any more questions, feel free to ask. Enjoy your day!
+
+
+if __name__ == "__main__":
+    asyncio.run(main())
