+++ python/samples/getting_started_with_agents/chat_completion/step3_kernel_function_strategies.py
@@ -4,10 +4,7 @@ import asyncio
 
 from semantic_kernel import Kernel
 from semantic_kernel.agents import AgentGroupChat, ChatCompletionAgent
-from semantic_kernel.agents.strategies import (
-    KernelFunctionSelectionStrategy,
-    KernelFunctionTerminationStrategy,
-)
+from semantic_kernel.agents.strategies import KernelFunctionSelectionStrategy, KernelFunctionTerminationStrategy
 from semantic_kernel.connectors.ai.open_ai import AzureChatCompletion
 from semantic_kernel.contents import AuthorRole, ChatMessageContent
 from semantic_kernel.functions import KernelFunctionFromPrompt
@@ -29,14 +26,29 @@ def _create_kernel_with_chat_completion(service_id: str) -> Kernel:
     return kernel
 
 
+REVIEWER_NAME = "ArtDirector"
+REVIEWER_INSTRUCTIONS = """
+You are an art director who has opinions about copywriting born of a love for David Ogilvy.
+The goal is to determine if the given copy is acceptable to print.
+If so, state that it is approved.
+If not, provide insight on how to refine suggested copy without example.
+"""
+
+COPYWRITER_NAME = "CopyWriter"
+COPYWRITER_INSTRUCTIONS = """
+You are a copywriter with ten years of experience and are known for brevity and a dry humor.
+The goal is to refine and decide on the single best copy as an expert in the field.
+Only provide a single proposal per response.
+You're laser focused on the goal at hand.
+Don't waste time with chit chat.
+Consider suggestions when refining an idea.
+"""
+
+TASK = "a slogan for a new line of electric cars."
+
+
 async def main():
-    REVIEWER_NAME = "ArtDirector"
-    REVIEWER_INSTRUCTIONS = """
-    You are an art director who has opinions about copywriting born of a love for David Ogilvy.
-    The goal is to determine if the given copy is acceptable to print.
-    If so, state that it is approved.
-    If not, provide insight on how to refine suggested copy without example.
-    """
+    # 1. Create the reviewer agent based on the chat completion service
     agent_reviewer = ChatCompletionAgent(
         service_id="artdirector",
         kernel=_create_kernel_with_chat_completion("artdirector"),
@@ -44,15 +56,7 @@ async def main():
         instructions=REVIEWER_INSTRUCTIONS,
     )
 
-    COPYWRITER_NAME = "CopyWriter"
-    COPYWRITER_INSTRUCTIONS = """
-    You are a copywriter with ten years of experience and are known for brevity and a dry humor.
-    The goal is to refine and decide on the single best copy as an expert in the field.
-    Only provide a single proposal per response.
-    You're laser focused on the goal at hand.
-    Don't waste time with chit chat.
-    Consider suggestions when refining an idea.
-    """
+    # 2. Create the copywriter agent based on the chat completion service
     agent_writer = ChatCompletionAgent(
         service_id="copywriter",
         kernel=_create_kernel_with_chat_completion("copywriter"),
@@ -60,6 +64,7 @@ async def main():
         instructions=COPYWRITER_INSTRUCTIONS,
     )
 
+    # 3. Create a Kernel Function to determine if the copy has been approved
     termination_function = KernelFunctionFromPrompt(
         function_name="termination",
         prompt="""
@@ -70,6 +75,7 @@ async def main():
         """,
     )
 
+    # 4. Create a Kernel Function to determine which agent should take the next turn
     selection_function = KernelFunctionFromPrompt(
         function_name="selection",
         prompt=f"""
@@ -91,6 +97,7 @@ async def main():
         """,
     )
 
+    # 5. Place the agents in a group chat with the custom termination and selection strategies
     chat = AgentGroupChat(
         agents=[agent_writer, agent_reviewer],
         termination_strategy=KernelFunctionTerminationStrategy(
@@ -110,15 +117,20 @@ async def main():
         ),
     )
 
-    input = "a slogan for a new line of electric cars."
+    # 6. Add the task as a message to the group chat
+    await chat.add_chat_message(ChatMessageContent(role=AuthorRole.USER, content=TASK))
+    print(f"# User: {TASK}")
 
-    await chat.add_chat_message(ChatMessageContent(role=AuthorRole.USER, content=input))
-    print(f"# User: '{input}'")
-
+    # 7. Invoke the chat
     async for content in chat.invoke():
-        print(f"# Agent - {content.name or '*'}: '{content.content}'")
+        print(f"# {content.name}: {content.content}")
 
-    print(f"# IS COMPLETE: {chat.is_complete}")
+    # Sample Output:
+    # User: a slogan for a new line of electric cars.
+    # CopyWriter: "Electrify your drive. Spare the gas, not the thrill."
+    # ArtDirector: This slogan captures the essence of electric cars but could use refinement to ...
+    # CopyWriter: "Go electric. Enjoy the thrill. Skip the gas."
+    # ArtDirector: Approved. This slogan is clear, concise, and effectively communicates the ...
 
 
 if __name__ == "__main__":
