+++ python/semantic_kernel/agents/chat_completion/chat_completion_agent.py
@@ -37,7 +37,7 @@ class ChatCompletionAgent(Agent):
     """
 
     service_id: str
-    channel_type: ClassVar[type[AgentChannel]] = ChatHistoryChannel
+    channel_type: ClassVar[type[AgentChannel] | None] = ChatHistoryChannel
 
     def __init__(
         self,
